Add download functionality
API: respond to GET request by providing all audio messages for a
conversation (sender/receiver to uniquely identify this) that the
client (sender) does not have. This is returned as a zip file as there
may be more than one audio file.
client: download zip of audio files and store them in the appropriate
directory.