Fixed a bug in matches API
Previously, the sender would only receive matches where they are the
user, and not when they are the match too!

This logic, if A has been matched with B, then if B queries it would
not identify a match with A.

However, this has been resolved, albeit in an inelegant way.