Save form data to the database
This is an inelegant way to store data (see: WTforms), but was faster
to implement as I did not have to separate the forms into another view
that would push relevant parts when necessary.

Note: removed matchmaking algorithm as it will be in a separate cron
job and is independent of views.