Added a bootloader for client.py
This script (along with a virtual environment and the codebase) must be
installed to the pi prior to dispatching the bespoke artefact to a
client.

This script would be used by a cron job (crontab) on the pi, which
would invoke it when the pi booted. This ensures that if the pi runs
out of battery, network issues, etc, then it will take off where it was
prior to the incident.