Wrote an initial skills matchmaking algorithm
Note: this algorithm is a proof-of-concept and therefore quite
simplistic in its approach. It looks for a single over lapping skill
between users who have not been previously matched.

Have to store on the route of the folder to make use of models and db
of the flask application code.